Surprise Change
Apparently, There was an update on the homework Wed-Thur that went
further in-depth on what was exactly required.

An attempt to implement the logic for testing, but do not have all the
necessary classes in place to perform this specific addition of
application testing.

Ran tests using the Log.D in previous versions to ensure that
everything worked properly.